fix: Corrected IAM and names to prevent length/perms issues (#2)
Changed name of resources to use substring of hash of vpc name since I ran into issues where the name length was to long. Additionally found that the IAM service account created for the cloudfunction was not actually referenced in the cloudfunction. After doing so, I found it needed additional permissions to access the pubsub topic.