fix: changed name to use random to avoid conflicts with same VPC names (#3)
Moved away from hash/substring in name of resources. This caused an issue where if the VPC name was the same, the bucket names would collide since they need to be globally unique. Decided to just use random string and standardized all the names to use similar convention. This prevents bucket naming collisions from occurring and allows you to deploy this multiple times within the same project for multiple VPCs. However, this could be tweaked further so you wouldn't need to duplicate the resources for multiple VPCs and work cross project but I won't do this as of now since it's not needed in my usecase.